requirements.txt
bleach==3.1.0
Django==2.2.5
django-crispy-forms==1.7.2
django-markdownify==0.8.0
Markdown==3.1.1
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
celery==4.3.0
redis==3.3.8
django-celery-results==1.1.2
factory_boy==2.12.0
